+++ src/ensembl/src/content/app/species-selector/state/speciesSelectorActions.ts
@@ -7,6 +7,7 @@ import { ActionCreator, Action } from 'redux';
 import { ThunkAction } from 'redux-thunk';
 import find from 'lodash/find';
 import get from 'lodash/get';
+import pickBy from 'lodash/pickBy';
 import apiService from 'src/services/api-service';
 
 import speciesSelectorStorageService from 'src/content/app/species-selector/services/species-selector-storage-service';
@@ -16,6 +17,7 @@ import buildAnalyticsObject from 'src/analyticsHelper';
 import {
   getCommittedSpecies,
   getCommittedSpeciesById,
+  getEnabledCommittedSpecies,
   getSelectedItem,
   getSearchText
 } from 'src/content/app/species-selector/state/speciesSelectorSelectors';
@@ -29,6 +31,8 @@ import {
   PopularSpecies,
   CommittedItem
 } from 'src/content/app/species-selector/types/species-search';
+
+import { getGenomeInfoById } from 'src/genome/genomeSelectors';
 
 import { CurrentItem } from './speciesSelectorState';
 
@@ -138,6 +142,30 @@ export const clearSelectedSearchResult = createStandardAction(
 //     dispatch(fetchStrainsAsyncActions.failure(error));
 //   }
 // };
+
+export const ensureSpeciesIsCommitted: ActionCreator<
+  ThunkAction<void, any, null, Action<string>>
+> = (genomeId: string) => async (dispatch, getState: () => RootState) => {
+  const state = getState();
+  const committedSpecies = getEnabledCommittedSpecies(state);
+  const genomeInfo = getGenomeInfoById(state, genomeId);
+  if (getCommittedSpeciesById(state, genomeId) || !genomeInfo) {
+    return;
+  }
+
+  const newCommittedSpecies = [
+    ...committedSpecies,
+    {
+      ...pickBy(genomeInfo, (value, key) => {
+        return key !== 'example_objects';
+      }),
+      isEnabled: true
+    }
+  ] as CommittedItem[];
+
+  dispatch(updateCommittedSpecies(newCommittedSpecies));
+  speciesSelectorStorageService.saveSelectedSpecies(newCommittedSpecies);
+};
 
 export const fetchAssemblies: ActionCreator<
   ThunkAction<void, any, null, Action<string>>
